requirements.txt
motor==3.0.0
pydantic==1.9.1
pymongo==4.2.0
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
